requirements.txt
flask==3.0.3
openai>=1.42.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
PyMySQL>=1.1.1
requests>=2.32.3
gunicorn